// src/app/components/models/user.ts
export interface UserInterface{
    email?: string;
    password?: string;
    cpassword?: string;
    photoURL?: any;
    displayName?: string;
    uid?: string;
    phoneNumber?: string;
    descripcion?: string;
    name?: string;
    id?: string;
    followers?: any;
    following?: any;
    scans?: any;
    peso?: number;
}